+++ comp/numberfespace.cpp
@@ -71,7 +71,7 @@ namespace ngcomp
       is_atomic_dof = true;
     }
 
-    virtual int GetNDof() const { return 1; }
+    virtual size_t GetNDof() const { return 1; }
 
     virtual FiniteElement & GetFE (ElementId ei, Allocator & lh) const
     {
